+++ persistence/src/test/java/edu/unc/lib/dl/persist/services/acl/ExpireEmbargoServiceIT.java
@@ -13,7 +13,7 @@
  * See the License for the specific language governing permissions and
  * limitations under the License.
  */
-package edu.unc.lib.dl.persist.services.edit;
+package edu.unc.lib.dl.persist.services.acl;
 
 import edu.unc.lib.dl.fcrepo4.AdminUnit;
 import edu.unc.lib.dl.fcrepo4.CollectionObject;
